Fix lint errors and disable nit-picky rules
Some of these rules are not really helpful in my
opinion and the style of the codebase varies a
little already anyway.

Bitwise operations are quite necessary for us
when we're trying to optimise some things or
implement hashing correctly.

"no-plusplus" really has no purpose, and the
same goes for "no-multi-assign".

"no-unused-vars" had no escape hatch without
excessive comments to disable it.

"import/prefer-default-export" probably leads
to slighly different names across some files.

"import/no-cycle" however can now be turned on
safely, which is nice
