+++ packages/styled-components/src/utils/hasher.js
@@ -1,4 +1,5 @@
 // @flow
+/* eslint-disable */
 
 import generateAlphabeticName from './generateAlphabeticName';
